Move transaction helper to shared location
