+++ tests/testapp/tests/sync/test_operations.py
@@ -1,16 +1,11 @@
 import json
-import threading
 import uuid
-from time import sleep
 
 import factory
 import mock
-import pytest
-from django.conf import settings
 from django.db import connection
 from django.test import override_settings
 from django.test import TestCase
-from django.test import TransactionTestCase
 from django.utils import timezone
 from facility_profile.models import ConditionalLog
 from facility_profile.models import Facility
@@ -36,7 +31,6 @@ from morango.sync.backends.utils import load_backend
 from morango.sync.context import LocalSessionContext
 from morango.sync.controller import MorangoProfileController
 from morango.sync.controller import SessionController
-from morango.sync.operations import _begin_transaction
 from morango.sync.operations import _dequeue_into_store
 from morango.sync.operations import _deserialize_from_store
 from morango.sync.operations import _queue_into_buffer_v1
@@ -76,93 +70,6 @@ def assertRecordsNotBuffered(records):
     for i in records:
         assert i.id not in buffer_ids
         assert i.id not in rmcb_ids
-
-
-def _concurrent_store_write(thread_event, store_id):
-    while not thread_event.is_set():
-        sleep(.1)
-    Store.objects.filter(id=store_id).delete()
-    connection.close()
-
-
-class TransactionIsolationTestCase(TransactionTestCase):
-    serialized_rollback = True
-
-    def _fixture_setup(self):
-        """Don't setup fixtures for this test case"""
-        pass
-
-    @override_settings(MORANGO_TEST_POSTGRESQL=False)
-    def test_begin_transaction(self):
-        """
-        Assert that we can start a transaction using our util and make some writes without
-        raising errors, specifically
-        """
-        # the utility we're testing here avoids setting the isolation level when this setting is True
-        # because tests usually run within their own transaction. By the time the isolation level
-        # is attempted to be set within a test, there have been reads and writes and the isolation
-        # cannot be changed
-        self.assertFalse(connection.in_atomic_block)
-        with _begin_transaction(None, isolated=True):
-            session = SyncSession.objects.create(
-                id=uuid.uuid4().hex,
-                profile="facilitydata",
-                last_activity_timestamp=timezone.now(),
-            )
-            transfer_session = TransferSession.objects.create(
-                id=uuid.uuid4().hex,
-                sync_session=session,
-                push=True,
-                last_activity_timestamp=timezone.now(),
-            )
-            create_buffer_and_store_dummy_data(transfer_session.id)
-
-        # manual cleanup
-        self.assertNotEqual(0, Store.objects.all().count())
-        # will cascade delete
-        SyncSession.objects.all().delete()
-        Store.objects.all().delete()
-
-    @pytest.mark.skipif(
-        not getattr(settings, "MORANGO_TEST_POSTGRESQL", False), reason="Not supported"
-    )
-    def test_transaction_isolation_handling(self):
-        from psycopg2.extensions import ISOLATION_LEVEL_REPEATABLE_READ
-
-        store = Store.objects.create(
-            id=uuid.uuid4().hex,
-            last_saved_instance=uuid.uuid4().hex,
-            last_saved_counter=1,
-            partition=uuid.uuid4().hex,
-            profile="facilitydata",
-            source_id="qqq",
-            model_name="qqq",
-        )
-
-        concurrent_event = threading.Event()
-        concurrent_thread = threading.Thread(
-            target=_concurrent_store_write,
-            args=(concurrent_event, store.id),
-        )
-        concurrent_thread.start()
-
-        # this test is only for postgres, but we don't want the code to know it's a test
-        with override_settings(MORANGO_TEST_POSTGRESQL=False):
-            try:
-                self.assertNotEqual(connection.connection.isolation_level, ISOLATION_LEVEL_REPEATABLE_READ)
-                with _begin_transaction(Filter(store.partition), isolated=True):
-                    self.assertEqual(connection.connection.isolation_level, ISOLATION_LEVEL_REPEATABLE_READ)
-                    s = Store.objects.get(id=store.id)
-                    concurrent_event.set()
-                    sleep(.2)
-                    s.last_saved_counter += 1
-                    s.save()
-                raise AssertionError("Didn't raise transactional error")
-            except Exception as e:
-                self.assertTrue(DBBackend._is_transaction_isolation_error(e))
-                self.assertNotEqual(connection.connection.isolation_level, ISOLATION_LEVEL_REPEATABLE_READ)
-            finally:
-                concurrent_thread.join(5)
 
 
 @override_settings(MORANGO_SERIALIZE_BEFORE_QUEUING=False, MORANGO_DISABLE_FSIC_V2_FORMAT=True)
